requirements.txt
langchain
langchain-astradb
langchain-huggingface
langchain-groq
langchain-community
datasets 
pypdf 
python_dotenv
pandas 
flask 
prometheus_client
